Перенес BaseGraph в GraphLight.Core
