+++ GraphLight.Core/BaseViewModel.cs
