+++ GraphLight.Core/Properties/Annotations.cs
